moved type alias
