+++ TentakelsAttacking2/Events/public/UIEvents.hpp
@@ -7,7 +7,7 @@
 #include "MainEvent.hpp"
 #include "HSoundType.hpp"
 #include "CustomRaylib.h"
-#include "CUIAlias.hpp"
+#include "UIAlias.hpp"
 #include <string>
 #include <functional>
