+++ TentakelsAttacking2/Game/public/GalaxyManager.h
@@ -4,7 +4,7 @@
 //
 
 #pragma once
-#include "CLogicAlias.hpp"
+#include "LogicAlias.hpp"
 #include <memory>
 
 class GameManager;
