+++ TentakelsAttacking2/Game/public/SpaceObject.h
@@ -4,7 +4,7 @@
 //
 
 #pragma once
-#include "CLogicAlias.hpp"
+#include "LogicAlias.hpp"
 #include "Vec2.hpp"
 #include <string>
 #include <memory>
