+++ TentakelsAttacking2/Helper/public/HGalaxy.h
@@ -6,7 +6,7 @@
 #pragma once
 #include <utility>
 #include <memory>
-#include <CLogicAlias.hpp>
+#include "LogicAlias.hpp"
 
 [[nodiscard]] std::pair<bool, SpaceObject_ty> TryGetTarget(
 	Fleet_ty_raw fleet, SpaceObject_ty const& target);
