+++ TentakelsAttacking2/TentakelsAttacking2.vcxproj
@@ -176,7 +176,7 @@
     <ClInclude Include="Game\public\SpaceObject.h" />
     <ClInclude Include="Game\public\TargetPoint.h" />
     <ClInclude Include="Helper\public\GameEventTypes.hpp" />
-    <ClInclude Include="Constants\public\CLogicAlias.hpp" />
+    <ClInclude Include="Helper\public\LogicAlias.hpp" />
     <ClInclude Include="Helper\public\HColors.h" />
     <ClInclude Include="Helper\public\HConcepts.hpp" />
     <ClInclude Include="Helper\public\HErrorLog.h" />
@@ -262,7 +262,7 @@
     <ClCompile Include="Constants\private\ConfigIO.cpp" />
     <ClCompile Include="Constants\private\CWindow.cpp" />
     <ClCompile Include="Constants\public\CFleet.hpp" />
-    <ClCompile Include="Constants\public\CUIAlias.hpp" />
+    <ClCompile Include="Helper\public\UIAlias.hpp" />
     <ClCompile Include="Game\private\GalaxyManager.cpp" />
     <ClCompile Include="Game\private\GameManager.cpp" />
     <ClCompile Include="Game\private\SpaceObject.cpp" />
