+++ TentakelsAttacking2/TentakelsAttacking2.vcxproj.filters
@@ -327,7 +327,10 @@
     <ClInclude Include="Helper\public\HGalaxy.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Constants\public\CLogicAlias.hpp">
+    <ClInclude Include="Helper\public\LogicAlias.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="UI\Elements\Default\public\LineDrag.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -563,7 +566,10 @@
     <ClCompile Include="Helper\private\HGalaxy.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Constants\public\CUIAlias.hpp">
+    <ClCompile Include="Helper\public\UIAlias.hpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="UI\Elements\Default\private\LineDrag.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
