+++ TentakelsAttacking2/UI/Elements/Galaxy/public/UIGalaxy.h
@@ -7,7 +7,7 @@
 #include "UIElement.hpp"
 #include "EventListener.hpp"
 #include "Focusable.h"
-#include "CLogicAlias.hpp"
+#include "LogicAlias.hpp"
 #include <functional>
 #include <memory>
 #include <vector>
